Pull classify into AddressManager
